# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для хранения и менеджмента пользовательских данных",
        "title": "GophKeeper API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/user/binary": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает бинарные данные",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "binary"
                ],
                "summary": "Создать бинарные данные",
                "parameters": [
                    {
                        "description": "Данные для сохранения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_brotigen23_goph-keeper_server_internal_dto_binarydto.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное создание"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Конфликт создания",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/binary/": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "binary"
                ],
                "summary": "Обновить существующую бинарную запись",
                "parameters": [
                    {
                        "description": "Данные для Обновления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_brotigen23_goph-keeper_server_internal_dto_binarydto.PutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет запись с входящим ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "binary"
                ],
                "summary": "Удалить существующую запись",
                "parameters": [
                    {
                        "description": "Данные с id записью",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Binary.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/binary/fetch": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список бинарных данных пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "binary"
                ],
                "summary": "Получить все бинарные данные",
                "responses": {
                    "200": {
                        "description": "Успешное выполнение"
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/cards": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Создать данные банковской карты",
                "parameters": [
                    {
                        "description": "Данные для сохранения",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Card.PostRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное создание"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Конфликт создания",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/cards/": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Обновить существующую карту",
                "parameters": [
                    {
                        "description": "Данные для Обновления",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Card.PutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное обновление"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет запись с входящим ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Удалить существующую запись",
                "parameters": [
                    {
                        "description": "Данные с id записью",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Card.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное удаление"
                    },
                    "400": {
                        "description": "Невалидные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/cards/fetch": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Получить все бинарные данные",
                "responses": {
                    "200": {
                        "description": "Успешное выполнение"
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Binary.DeleteRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "Card.DeleteRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "Card.PostRequest": {
            "type": "object",
            "properties": {
                "cardholder_name": {
                    "type": "string",
                    "example": "IVAN IVANOV"
                },
                "cvv": {
                    "type": "string",
                    "example": "123"
                },
                "expire": {
                    "type": "string",
                    "example": "29.01.2033"
                },
                "metadata": {
                    "type": "string",
                    "example": "metadata"
                },
                "number": {
                    "type": "string",
                    "example": "1234 5678 9123 456"
                }
            }
        },
        "Card.PutRequest": {
            "type": "object",
            "properties": {
                "cardholder_name": {
                    "type": "string",
                    "example": "IVAN IVANOV"
                },
                "cvv": {
                    "type": "string",
                    "example": "123"
                },
                "expire": {
                    "type": "string",
                    "example": "29.01.2033"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "metadata": {
                    "type": "string",
                    "example": "metadata"
                },
                "number": {
                    "type": "string",
                    "example": "1234 5678 9123 456"
                }
            }
        },
...
        "github_com_brotigen23_goph-keeper_server_internal_dto_binarydto.PostRequest": {
            "type": "object"
        },
        "github_com_brotigen23_goph-keeper_server_internal_dto_binarydto.PutRequest": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Введите \"Bearer \u003cJWT\u003e\", где JWT - ваш access токен",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}